# docs/services/gateway/api/public.openapi.yaml
openapi: 3.0.3
info:
  title: Power Knowledge Platform Gateway API
  version: 0.1.0
  license:
    name: Unspecified
    url: https://github.com/Sakayori-Iroha-168/Software_Teamwork
  description: |
    Public caller-facing API contract for the gateway service.
    All stable public and service-to-service HTTP APIs must use RESTful resource-oriented paths.
    This revision treats gateway health checks, user/session,
    knowledge-owned knowledge base, document upload, processing, content, chunk, and retrieval routes,
    document-owned report generation and report settings routes, qa-owned intelligent question answering
    routes, admin-facing runtime model/parser configuration routes, admin user
    management, personal profile, required password-change routes, and gateway-owned
    app version freshness routes as active contracts. Both admin-overview and
    admin-metrics are active gateway contracts for the administration dashboard.
...
servers:
- url: /
  description: Gateway root
security:
- bearerAuth: []
...
paths:
...
  /api/v1/report-types:
    get:
      tags:
      - report-generation
      summary: List report types
      operationId: listReportTypes
      x-owner-service: document
      responses:
        '200':
          description: Report types.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTypeListResponse'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/report-templates:
    get:
      tags:
      - report-generation
      summary: List report templates
      operationId: listReportTemplates
      x-owner-service: document
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - name: reportType
        in: query
        schema:
          type: string
      - name: enabled
        in: query
        schema:
          type: boolean
      responses:
        '200':
          description: Report templates.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplateListResponse'
        '400':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - report-generation
      summary: Create report template
      operationId: createReportTemplate
      x-owner-service: document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateReportTemplateRequest'
      responses:
        '201':
          description: Report template created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplateResponse'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/report-templates/{reportTemplateId}:
    parameters:
    - $ref: '#/components/parameters/ReportTemplateId'
    get:
      tags:
      - report-generation
      summary: Get report template
      operationId: getReportTemplate
      x-owner-service: document
      responses:
        '200':
          description: Report template.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplateResponse'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - report-generation
      summary: Update report template
      operationId: updateReportTemplate
      x-owner-service: document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportTemplateRequest'
      responses:
        '200':
          description: Report template updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplateResponse'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags:
      - report-generation
      summary: Delete report template
      operationId: deleteReportTemplate
      x-owner-service: document
      description: Prefer disabling or soft deleting templates that have been used by reports.
      responses:
        '204':
          description: Report template deleted.
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/report-templates/{reportTemplateId}/structure:
    parameters:
    - $ref: '#/components/parameters/ReportTemplateId'
    get:
      tags:
      - report-generation
      summary: Get report template structure
      operationId: getReportTemplateStructure
      x-owner-service: document
      responses:
        '200':
          description: Report template structure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplateStructureResponse'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - report-generation
      summary: Update report template structure
      operationId: updateReportTemplateStructure
      x-owner-service: document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportTemplateStructureRequest'
      responses:
        '200':
          description: Report template structure updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplateStructureResponse'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/report-materials:
    get:
      tags:
      - report-generation
      summary: List report materials
      operationId: listReportMaterials
      x-owner-service: document
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - name: category
        in: query
        schema:
          type: string
      - name: enabled
        in: query
        schema:
          type: boolean
      responses:
        '200':
          description: Report materials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportMaterialListResponse'
        '400':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - report-generation
      summary: Create report material
      operationId: createReportMaterial
      x-owner-service: document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateReportMaterialRequest'
      responses:
        '201':
          description: Report material created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportMaterialResponse'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/report-materials/{materialId}:
    parameters:
    - $ref: '#/components/parameters/MaterialId'
    get:
      tags:
      - report-generation
      summary: Get report material
      operationId: getReportMaterial
      x-owner-service: document
      responses:
        '200':
          description: Report material.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportMaterialResponse'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags:
      - report-generation
      summary: Delete report material
      operationId: deleteReportMaterial
      x-owner-service: document
      description: Prefer soft deletion when material has been referenced by generation jobs.
      responses:
        '204':
          description: Report material deleted.
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
  /api/v1/reports/{reportId}/jobs:
    parameters:
    - $ref: '#/components/parameters/ReportId'
    get:
      tags:
      - report-generation
      summary: List report jobs
      operationId: listReportJobs
      x-owner-service: document
      responses:
        '200':
          description: Report jobs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobListResponse'
        '404':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - report-generation
      summary: Create report job
      operationId: createReportJob
      x-owner-service: document
      description: Creates long-running work for outline creation, content creation, section regeneration, or report file creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportJobRequest'
      responses:
        '202':
          description: Report job accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobResponse'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/report-jobs/{jobId}:
    parameters:
    - $ref: '#/components/parameters/JobId'
    get:
      tags:
      - report-generation
      summary: Get report job
      operationId: getReportJob
      x-owner-service: document
      responses:
        '200':
          description: Report job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobResponse'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - report-generation
      summary: Update report job
      operationId: updateReportJob
      x-owner-service: document
      description: Updates a report job state. The initial supported update is canceling pending or running generation work.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportJobRequest'
      responses:
        '200':
          description: Report job updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobResponse'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/report-jobs/{jobId}/attempts:
    parameters:
    - $ref: '#/components/parameters/JobId'
    get:
      tags:
      - report-generation
      summary: List report job attempts
      operationId: listReportJobAttempts
      x-owner-service: document
      responses:
        '200':
          description: Report job attempts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobAttemptListResponse'
        '404':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - report-generation
      summary: Create report job attempt
      operationId: createReportJobAttempt
      x-owner-service: document
      description: Creates a new attempt for an existing failed or interrupted report job.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportJobAttemptRequest'
      responses:
        '202':
          description: Report job attempt accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobAttemptResponse'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/reports/{reportId}/events:
    parameters:
    - $ref: '#/components/parameters/ReportId'
    get:
      tags:
      - report-generation
      summary: List report events
      operationId: listReportEvents
      x-owner-service: document
      description: Returns report workflow events for polling. A future SSE contract must be added explicitly before streaming clients rely on it.
      responses:
        '200':
          description: Report events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportEventListResponse'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/report-files:
    get:
      tags:
      - report-generation
      summary: List report files
      operationId: listReportFiles
      x-owner-service: document
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - name: reportId
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Report files.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportFileListResponse'
        '400':
          $ref: '#/components/responses/Error'
    post:
      tags:
      - report-generation
      summary: Create report file
      operationId: createReportFile
      x-owner-service: document
      description: Creates generated report files such as DOCX without exposing storage internals to frontend clients. Document service owns report-file business metadata and should use file service for underlying object storage/content access instead of duplicating public file semantics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportFileRequest'
      responses:
        '202':
          description: Report file accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportFileResponse'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/report-files/{reportFileId}:
    parameters:
    - $ref: '#/components/parameters/ReportFileId'
    get:
      tags:
      - report-generation
      summary: Get report file
      operationId: getReportFile
      x-owner-service: document
      responses:
        '200':
          description: Report file metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportFileResponse'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/report-files/{reportFileId}/content:
    parameters:
    - $ref: '#/components/parameters/ReportFileId'
    get:
      tags:
      - report-generation
      summary: Get report file content
      operationId: getReportFileContent
      x-owner-service: document
      description: Returns generated file content. Error responses still use the standard gateway error envelope.
      responses:
        '200':
          description: Generated report file stream.
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/report-statistics/overview:
    get:
      tags:
      - report-generation
      summary: Get report statistics overview
      operationId: getReportStatisticsOverview
      x-owner-service: document
      responses:
        '200':
          description: Report statistics overview.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportStatisticsOverviewResponse'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/report-statistics/daily:
    get:
      tags:
      - report-generation
      summary: List daily report statistics
      operationId: listDailyReportStatistics
      x-owner-service: document
      parameters:
      - name: days
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 366
          default: 30
      responses:
        '200':
          description: Daily report statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDailyStatisticsResponse'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/report-operation-logs:
    get:
      tags:
      - report-generation
      summary: List report operation logs
      operationId: listReportOperationLogs
      x-owner-service: document
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - name: targetType
        in: query
        schema:
          type: string
      - name: targetId
        in: query
        schema:
          type: string
      - name: operationType
        in: query
        schema:
          type: string
      - name: requestId
        in: query
        schema:
          type: string
      - name: requestSource
        in: query
        schema:
          type: string
      - name: toolName
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Report operation logs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOperationLogListResponse'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/report-settings:
    get:
      tags:
      - report-generation
      summary: Get report generation settings
      operationId: getReportSettings
      x-owner-service: document
      x-gateway-role: Public admin entrypoint and admin authorization. Document service owns report generation settings; AI Gateway owns model profile credentials.
      description: Returns admin-visible report generation settings. Provider base URLs and API keys remain owned by AI Gateway model profiles and are never returned here.
      responses:
        '200':
          description: Report generation settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSettingsResponse'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - report-generation
      summary: Update report generation settings
      operationId: updateReportSettings
      x-owner-service: document
      x-gateway-role: Public admin entrypoint and admin authorization. Document service validates referenced templates and AI Gateway model profile ids.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportSettingsRequest'
      responses:
        '200':
          description: Report generation settings updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSettingsUpdateResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token issued by auth and cached by gateway through a token hash; callers must not treat it as a JWT.
  parameters:
...
    ReportTemplateId:
      name: reportTemplateId
      in: path
      required: true
      schema:
        type: string
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ReportTypeListResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportType'
        requestId:
          type: string
    CreateReportTemplateRequest:
      type: object
      required:
      - file
      - templateName
      - reportType
      properties:
        file:
          type: string
          format: binary
        templateName:
          type: string
        reportType:
          type: string
        description:
          type: string
    UpdateReportTemplateRequest:
      type: object
      properties:
        templateName:
          type: string
        description:
          type: string
        enabled:
          type: boolean
...
    ReportTemplateResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          $ref: '#/components/schemas/ReportTemplate'
        requestId:
          type: string
    ReportTemplateListResponse:
      type: object
      required:
      - data
      - page
      - requestId
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportTemplate'
        page:
          $ref: '#/components/schemas/PageMeta'
        requestId:
          type: string
...
    ReportTemplateStructureRequest:
      $ref: '#/components/schemas/ReportTemplateStructure'
    ReportTemplateStructureResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          $ref: '#/components/schemas/ReportTemplateStructure'
        requestId:
          type: string
    CreateReportMaterialRequest:
      type: object
      required:
      - file
      - materialName
      - materialType
      properties:
        file:
          type: string
          format: binary
        materialName:
          type: string
        materialType:
          type: string
        category:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
...
    ReportMaterialResponse:
      type: object
      required:
      - data
      - requestId
      properties:
        data:
          $ref: '#/components/schemas/ReportMaterial'
        requestId:
          type: string
    ReportMaterialListResponse:
      type: object
      required:
      - data
      - page
      - requestId
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportMaterial'
        page:
          $ref: '#/components/schemas/PageMeta'
        requestId:
          type: string
...